feat: Add confirmation steps to payroll and reset commands
This commit introduces confirmation steps for the `/payroll` and `/reset` administrative commands to prevent accidental data loss.

The `/payroll` command now includes a required `confirm: True` boolean parameter. The command will only proceed if this parameter is provided and set to `True`. The command registration in `bot.py` has been updated to reflect this new required parameter. The payroll receipt embed now also includes a detailed list of users paid and the amounts they received.

The `/reset` command now has a two-step confirmation process. It first requires a `confirm: True` parameter. If confirmed, it then presents the administrator with a secondary confirmation message containing "Confirm" and "Cancel" buttons. The irreversible data reset will only occur if the "Confirm" button is clicked.

A reusable `ConfirmView` class has been added to the `views/` directory to handle the button-based confirmation logic. The view now defers the interaction by calling `interaction.response.defer()` to prevent timeouts on potentially long-running operations.

The `pay_all_pending_melange` function in `database_orm.py` has been corrected to use a single transaction, ensuring the list of paid users is correctly populated and returned.

The test suite has been updated and restructured to follow project conventions. New, dedicated test files (`test_payroll_command.py`, `test_reset_command.py`) have been added for the modified commands, and general smoke tests have been preserved in `test_commands.py`.
